+++ parac-build.py
@@ -1,15 +1,16 @@
 # coding=utf-8
-""" Script for building the compiler as an executable """
+"""
+CLI Script for building the compiler and generating the binaries that may be
+used.
+"""
+import argparse
 import json
 import os
 import platform
-import re
 import shutil
-import sys
 import time
-from distutils.dir_util import copy_tree
 from pathlib import Path
-from typing import List, Dict, Optional, Union, NoReturn, Tuple
+from typing import List, Literal
 
 import PyInstaller.__main__
 import requests
@@ -20,39 +21,26 @@ COMPATIBLE_VERSION: str = "v0.1.dev6"
 # Input regex
 IN_REGEX: str = "(--[A-z0-9]+=[A-z0-9]+)"
 
-# Run Constants
+# Path Constants
 BASE_PATH: Path = Path(os.getcwd())
 DIST_PATH: Path = BASE_PATH / "dist"
 BUILD_PATH: Path = BASE_PATH / "build"
 ENTRY_PATH: Path = BASE_PATH / "dummy-entry.py"
 EXAMPLE_PATH: Path = BASE_PATH / "example"
 ICON_PATH: Path = BASE_PATH / "img" / "parac.ico"
+
+# C Compiler and PBL Config
 C_COMPILER = "gcc"
 C_LIB_IDENTIFIER = "libpbl.a"
-LIBPBL = f"https://github.com/Para-C/Para-C-Base-Library/releases/download/{COMPATIBLE_VERSION}/{C_LIB_IDENTIFIER}"
+LIB_PBL = f"https://github.com/Para-C/Para-C-Base-Library/releases/download/{COMPATIBLE_VERSION}/{C_LIB_IDENTIFIER}"
 C_LIB_TEMP_DESTINATION = BASE_PATH / "tmp" / C_LIB_IDENTIFIER
+
+# Global destination paths
+POSIX_GLOBAL_DEST: Path = Path("/usr/local/bin/Para-C")
+NT_GLOBAL_DEST: Path = Path("C:\\Program Files (x86)\\Para-C\\")
+
+# Target system
 TARGET: str = platform.system()
-IS_64_BIT: bool = sys.maxsize > 2 ** 32
-
-# Configuration Defaults
-ACTION: str = "normal"
-
-# Possible Configuration
-POSSIBLE_ACTION_CONF: Tuple[str, str, str] = (
-    "normal", "pbl", "compiler-standalone"
-)
-
-# Etc.
-ACTIONS_THAT_REQUIRE_PBL: Tuple[str, str] = (
-    "normal", "pbl"
-)
-
-# Run Configuration - default values are passed onto the dictionary
-CONFIG: Dict[str, Optional[str]] = {
-    "ACTION": ACTION,
-    "TARGET": TARGET,
-    "LIBPBL": LIBPBL
-}
 
 # Required additional data files that have to be added
 COPY_FILES: List[Path] = [
@@ -79,14 +67,13 @@ HIDDEN_IMPORT = ["parac_ext_cli", "parac"]
 
 def create_bin_config(dest_dir: Path) -> None:
     """
-    Creates the binaries config file - Uses empty-bin-config.json as reference
+    Creates the binaries' config file
     """
     conf = {
         "c-compiler": C_COMPILER,
-        "target": CONFIG["TARGET"],
-        "mode": CONFIG["ACTION"],
+        "target": TARGET,
         "version": COMPATIBLE_VERSION,
-        "build-time": time.time()  # unix timestamp
+        "build-time": time.time()  # Unix timestamp
     }
     with open(dest_dir / "bin-config.json", "w+") as f:
         f.write(json.dumps(conf, indent=4))
@@ -109,33 +96,6 @@ def parse_avoid_and_hidden_imports() -> None:
     HIDDEN_IMPORT = _
 
 
-def validate_action() -> Union[None, NoReturn]:
-    """ Validates whether the action is valid """
-    if CONFIG["ACTION"] not in POSSIBLE_ACTION_CONF:
-        raise ValueError(
-            f"Invalid argument 'ACTION' with value '{CONFIG['ACTION']}'"
-        )
-    return
-
-
-def parse_input() -> None:
-    """ Parses the input and updates the global constants """
-    args = sys.argv[1:]
-
-    if not all(map(lambda n: bool(re.findall(IN_REGEX, n)), args)):
-        raise ValueError("Invalid input! Check Syntax: '--ARG=VALUE'")
-
-    for in_arg in args:
-        in_arg: str = in_arg[2:]
-        name, value = tuple(in_arg.split('='))
-        if name not in CONFIG.keys():
-            raise ValueError("Invalid argument name!")
-
-        CONFIG[name] = value
-
-    validate_action()
-
-
 def cleanup_and_create_tmp_folder() -> None:
     """
     Creates the temp folder for this run
@@ -153,85 +113,196 @@ def cleanup_and_remove_tmp_folder() -> None:
         shutil.rmtree(tmp)
 
 
-def fetch_pbl_build() -> None:
-    """ Fetches the pbl build """
-    # if there is an independent built of libpbl, then use that
-    if CONFIG["LIBPBL"] != LIBPBL:
-        if not (p := Path(CONFIG["LIBPBL"]).resolve()).exists():
-            raise ValueError(
-                f"Invalid argument 'LIBPBL' with value '{CONFIG['LIBPBL']}'"
-            )
-        shutil.copy(p, C_LIB_TEMP_DESTINATION)
-        return
+def _download_url(url: str) -> None:
+    """
+    Downloads the specified url
 
-    # fetching the file and writing it to the temp destination
-    r = requests.get(CONFIG["LIBPBL"], allow_redirects=True)
+    :raises RuntimeError: If the request failed to execute
+    """
+    # Fetching the file and writing it to the temp destination
+    r = requests.get(url, allow_redirects=True)
     if r.status_code != 200:
         raise RuntimeError(
-            f"Failed to execute request! Received '{r.status_code}' "
-            f"for endpoint {LIBPBL}"
+            f"Failed to execute request! Received "
+            f"'{r.status_code}' for endpoint {LIB_PBL}"
         )
 
     with open(C_LIB_TEMP_DESTINATION, "wb+") as f:
         f.write(r.content)
 
 
-def create_parac_modules(output_type: str) -> None:
-    """ Creates the required parac modules for the compiler """
-    # copying the generated files to the tmp directory
+def fetch_pbl_build(lib_pbl: str, allow_download: bool = False) -> None:
+    """
+    Fetches the pbl build and stores it in the 'C_LIB_TEMP_DESTINATION'
+
+    :param lib_pbl: The required path to the 'libpbl.a'. Can be a URL
+    :param allow_download: If set to True, the lib_pbl parameter will be taken
+     as a URL to download. (Only if the lib_pbl path wasn't found locally \
+     though)
+    """
+    # If there is an independent build of libpbl, then try to fetch that
+    if lib_pbl != LIB_PBL:
+        # If the path does not exist and allow_download is false
+        if not (p := Path(lib_pbl).resolve()).exists():
+            if allow_download is False:
+                raise ValueError(
+                    f"Could not find '--libpbl' with value '{lib_pbl}'"
+                )
+            else:
+                _download_url(lib_pbl)
+
+        shutil.copy(p, C_LIB_TEMP_DESTINATION)
+        return
+
+    _download_url(lib_pbl)
+
+
+def install_parac_module(output_type: Literal["dist", "build"]) -> None:
+    """
+    Creates the required parac module for the compiler
+
+    :param output_type: The output_type, where the data should be fetched from
+     and the environment installed.
+    """
+    # Copying the generated files to the tmp directory
     shutil.move(
         BASE_PATH / output_type / "parac",
         str(origin := BASE_PATH / "tmp" / output_type)
     )
 
-    # destination where the files should be copied to
+    # The destination where the files should be copied to
     destination = (BASE_PATH / output_type / "parac").resolve()
+    # The destination path of the compiled files
+    bin_path: Path = (destination / "bin").resolve()
 
-    # removing any previous data
+    # Removing any previous data
     if os.path.exists(destination):
         shutil.rmtree(str(destination))
 
-    bin_path: Path = (destination / "bin").resolve()
-    avoid = ["bin"]
-
-    # creating the binary directory
+    # Creating the binary directory
     os.makedirs(str(bin_path), exist_ok=True)
 
-    # copying the lib item if the mode says we have to
-    if CONFIG["ACTION"] in ACTIONS_THAT_REQUIRE_PBL:
-        shutil.copy(
-            C_LIB_TEMP_DESTINATION,
-            bin_path / C_LIB_IDENTIFIER
-        )
+    # Copying the required library with the c-headers
+    shutil.copy(
+        C_LIB_TEMP_DESTINATION,
+        bin_path / C_LIB_IDENTIFIER
+    )
 
     # All items are going to be copied to the bin path, since the output of
     # pyinstaller will already contain the binary as needed, so the rest of the
     # items will be from the Para-C module and managed by us
     for entry in os.scandir(origin):
         entry: os.DirEntry
-        # Avoid directories that are specifically marked as 'avoid'
-        if not (entry.name in avoid and entry.is_dir()):
+        # Avoid directories that are named "bin" and are a directory
+        if not (entry.name == "bin" and entry.is_dir()):
             shutil.move(entry.path, bin_path)
 
+    # Copies all extra specified files into the destination
+    # e.g. images, text files ...
     for entry in COPY_FILES:
         entry: Path
         shutil.copy(str(entry.resolve()), destination)
 
-    copy_tree(
+    # Copies the examples into the destination folder
+    shutil.copytree(
         str(EXAMPLE_PATH.resolve()),
         str((destination / "example").resolve())
     )
 
+    # Creates the binary configuration file
     create_bin_config(bin_path)
 
-    # remove temp for this build mode
+    # Remove temp files after finishing and having moved everything to its
+    # proper destination
     shutil.rmtree(origin)
 
 
+def install_global(optional_path: str):
+    """
+    Moves the content from the 'dist' folder into the OS-specific folder for
+    program files.
+
+    This will automatically fetch the proper destination based on the OS:
+    - Linux: /usr/local/bin
+    - MacOS: /usr/local/bin
+    - Windows: C:\Program Files (x86)\
+
+    :param optional_path: The optional path that may be specified where the
+     dest files should be copied to.
+    """
+    try:
+        path: Path
+        origin = (DIST_PATH / "parac").resolve()
+        if optional_path:
+            path: Path = Path(optional_path).resolve()
+            shutil.copytree(
+                src=origin,
+                dst=str(path)
+            )
+        else:
+            # Linux, MacOS, ...
+            if os.name == 'posix':
+                path: Path = Path(POSIX_GLOBAL_DEST).resolve()
+                shutil.copytree(
+                    src=origin,
+                    dst=str(path)
+                )
+            # Windows
+            elif os.name == 'nt':
+                path: Path = Path(NT_GLOBAL_DEST).resolve()
+                shutil.copytree(
+                    src=origin,
+                    dst=str(path)
+                )
+            else:
+                raise RuntimeError(
+                    f"Unsupported operating system (os.name): {os.name}"
+                )
+        print(
+            f"\nInstalled Para-C {COMPATIBLE_VERSION} globally in: {path}\n\n"
+            "For info on how to add the item to your path/create a bash alias,"
+            " as well as general info, go here: "
+            "https://para-c.readthedocs.io/en/latest/installation.html"
+        )
+    except IOError as e:
+        raise RuntimeError(
+            "Failed to install parac globally. Likely missing permissions"
+        ) from e
+
+
 if __name__ == "__main__":
-    parse_input()
+    parser = argparse.ArgumentParser(
+        description='Para-C Build Script'
+    )
+    parser.add_argument(
+        "--libpbl", nargs=1, default=LIB_PBL, type=str, required=False,
+        help="The destination path to the 'libpbl.a' file. If this is a url "
+             "that should be downloaded, please specify '--allow-download' "
+             "(Defaults to default mirror on GitHub)."
+    )
+    parser.add_argument(
+        "--allow-download", action="store_true", default=False, required=False,
+        help="If the 'libpbl' parameter is a URL that may be downloaded, then "
+             "this flag has to be set to avoid an error while processing!"
+    )
+    parser.add_argument(
+        "--install-global", action="store_true", default=False,
+        required=False,
+        help="If this is specified, the script will attempt to directly "
+             "install parac into your binary folder, so you can access it "
+             "directly after this script finished. (Create Alias/Add to PATH "
+             "yourself)."
+    )
+    parser.add_argument(
+        "--g-dest", type=str, default=None, required=False,
+        help="Specifies the global destination folder where parac should be "
+             "moved to. This is only valid when '--install-global' is also "
+             "specified otherwise the value is just ignored. "
+    )
+    args = parser.parse_args()
+
     cleanup_and_create_tmp_folder()
-    fetch_pbl_build()
+    fetch_pbl_build(args.libpbl, args.allow_download)
 
     # creating the output directories
     for i in (DIST_PATH, BUILD_PATH):
@@ -266,10 +337,16 @@ if __name__ == "__main__":
         *HIDDEN_IMPORT
     ]
 
-    # running pyinstaller - the output will appear in ./dist and ./build
+    # Running pyinstaller - the output will appear in ./dist and ./build
     PyInstaller.__main__.run(run_config)
 
-    create_parac_modules("dist")
-    create_parac_modules("build")
+    # Installs from the sources the proper para-c environment
+    install_parac_module("dist")
+    install_parac_module("build")
 
+    # If this is specified try to download the specified
+    if args.install_global:
+        install_global(args.g_dest)
+
+    # Cleanups the tmp files generated while running
     cleanup_and_remove_tmp_folder()